move all screen pixel transfer functions to general stubs
