+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D9/DX9draw.cpp
@@ -22,14 +22,6 @@
 #include "Graphics_Systems/General/GScolors.h"
 #include "Graphics_Systems/General/GScolor_macros.h"
 
-#include "Universal_System/roomsystem.h"
-
-#include <vector>
-#include <math.h>
-#include <stdio.h>
-
-using std::vector;
-
 using namespace enigma::dx9;
 
 namespace enigma_user {
@@ -54,93 +46,6 @@ int draw_get_msaa_maxlevel()
 bool draw_get_msaa_supported()
 {
   return false; //TODO: implement
-}
-
-} // namespace enigma_user
-
-//#include <endian.h>
-//TODO: Though serprex, the author of the function below, never included endian.h,
-//   // Doing so is necessary for the function to work at its peak.
-//   // When ENIGMA generates configuration files, one should be included here.
-
-namespace enigma_user {
-
-int draw_getpixel(int x, int y)
-{
-    if (view_enabled)
-    {
-        x = x + enigma_user::view_xview[enigma_user::view_current];
-        y = (y + enigma_user::view_yview[enigma_user::view_current]) - 1;
-        if (x < 0) x = 0;
-        if (y < 0) y = 0;
-        if (x > enigma_user::view_wview[enigma_user::view_current] || y > enigma_user::view_hview[enigma_user::view_current]) return 0;
-    } else {
-        if (x < 0) x = 0;
-        if (y < 0) y = 0;
-        if (x > enigma_user::room_width || y > enigma_user::room_height) return 0;
-    }
-
-	draw_batch_flush(batch_flush_deferred);
-
-	LPDIRECT3DSURFACE9 pBackBuffer;
-	LPDIRECT3DSURFACE9 pDestBuffer;
-	d3ddev->GetBackBuffer(0, 0, D3DBACKBUFFER_TYPE_MONO, &pBackBuffer);
-	D3DSURFACE_DESC desc;
-	pBackBuffer->GetDesc(&desc);
-
-	d3ddev->CreateOffscreenPlainSurface( desc.Width, desc.Height, desc.Format, D3DPOOL_SYSTEMMEM, &pDestBuffer, NULL );
-	d3ddev->GetRenderTargetData(pBackBuffer, pDestBuffer);
-
-	D3DLOCKED_RECT rect;
-
-	pDestBuffer->LockRect(&rect, NULL, D3DLOCK_READONLY);
-	unsigned char* bitmap = static_cast<unsigned char*>(rect.pBits);
-	unsigned offset = y * rect.Pitch + x * 4;
-	int ret = bitmap[offset + 2] | (bitmap[offset + 1] << 8) | (bitmap[offset + 0] << 16);
-	pDestBuffer->UnlockRect();
-
-	pBackBuffer->Release();
-	pDestBuffer->Release();
-
-	return ret;
-}
-
-int draw_getpixel_ext(int x, int y)
-{
-    if (view_enabled)
-    {
-        x = x + enigma_user::view_xview[enigma_user::view_current];
-        y = (y + enigma_user::view_yview[enigma_user::view_current]) - 1;
-        if (x < 0) x = 0;
-        if (y < 0) y = 0;
-        if (x > enigma_user::view_wview[enigma_user::view_current] || y > enigma_user::view_hview[enigma_user::view_current]) return 0;
-    } else {
-        if (x < 0) x = 0;
-        if (y < 0) y = 0;
-        if (x > enigma_user::room_width || y > enigma_user::room_height) return 0;
-    }
-
-	draw_batch_flush(batch_flush_deferred);
-
-	LPDIRECT3DSURFACE9 pBackBuffer;
-	LPDIRECT3DSURFACE9 pDestBuffer;
-	d3ddev->GetBackBuffer(0, 0, D3DBACKBUFFER_TYPE_MONO, &pBackBuffer);
-	D3DSURFACE_DESC desc;
-	pBackBuffer->GetDesc(&desc);
-	d3ddev->CreateOffscreenPlainSurface( desc.Width, desc.Height, desc.Format, D3DPOOL_SYSTEMMEM, &pDestBuffer, NULL );
-	d3ddev->GetRenderTargetData(pBackBuffer, pDestBuffer);
-
-	D3DLOCKED_RECT rect;
-
-	pDestBuffer->LockRect(&rect, NULL, D3DLOCK_READONLY);
-	unsigned char* bitmap = static_cast<unsigned char*>(rect.pBits);
-	unsigned offset = y * rect.Pitch + x * 4;
-	int ret = bitmap[offset + 2] | (bitmap[offset + 1] << 8) | (bitmap[offset + 0] << 16) | (bitmap[offset + 3] << 24);
-	pDestBuffer->UnlockRect();
-
-	pBackBuffer->Release();
-	pDestBuffer->Release();
-	return ret;
 }
 
 } // namespace enigma_user
