+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D9/DX9screen.cpp
@@ -17,22 +17,13 @@
 
 #include "Direct3D9Headers.h"
 #include "Graphics_Systems/General/GSscreen.h"
-#include "Graphics_Systems/General/GSbackground.h"
-#include "Graphics_Systems/General/GSsprite.h"
 #include "Graphics_Systems/General/GSprimitives.h"
-#include "Graphics_Systems/General/GScolors.h"
 
-#include "Universal_System/image_formats.h"
-#include "Universal_System/nlpo2.h"
-#include "Universal_System/Resources/background_internal.h"
-#include "Universal_System/Resources/sprites_internal.h"
 #include "Universal_System/roomsystem.h"
 
 #include "Platforms/General/PFwindow.h"
 
-using namespace enigma;
 using namespace enigma::dx9;
-using namespace std;
 
 namespace enigma {
 
@@ -56,82 +47,6 @@ void screen_set_viewport(gs_scalar x, gs_scalar y, gs_scalar width, gs_scalar he
   sy = (window_get_height() - window_get_region_height_scaled()) / 2;
 	D3DVIEWPORT9 pViewport = { (DWORD)(sx + x), (DWORD)(sy + y), (DWORD)width, (DWORD)height, 0, 1.0f };
 	d3ddev->SetViewport(&pViewport);
-}
-
-int screen_save(string filename) //Assumes native integers are little endian
-{
-  draw_batch_flush(batch_flush_deferred);
-
-	string ext = enigma::image_get_format(filename);
-
-	LPDIRECT3DSURFACE9 pBackBuffer;
-	LPDIRECT3DSURFACE9 pDestBuffer;
-	d3ddev->GetBackBuffer(0, 0, D3DBACKBUFFER_TYPE_MONO, &pBackBuffer);
-	D3DSURFACE_DESC desc;
-	pBackBuffer->GetDesc(&desc);
-
-	d3ddev->CreateOffscreenPlainSurface( desc.Width, desc.Height, desc.Format, D3DPOOL_SYSTEMMEM, &pDestBuffer, NULL );
-	d3ddev->GetRenderTargetData(pBackBuffer, pDestBuffer);
-
-	D3DLOCKED_RECT rect;
-
-	pDestBuffer->LockRect(&rect, NULL, D3DLOCK_READONLY);
-	unsigned char* bitmap = static_cast<unsigned char*>(rect.pBits);
-	pDestBuffer->UnlockRect();
-
-	int ret = image_save(filename, bitmap, desc.Width, desc.Height, desc.Width, desc.Height, false);
-
-  pBackBuffer->Release();
-	pDestBuffer->Release();
-
-	//delete[] bitmap; <- no need cause RAII
-	return ret;
-}
-
-int screen_save_part(string filename,unsigned x,unsigned y,unsigned w,unsigned h) //Assumes native integers are little endian
-{
-  draw_batch_flush(batch_flush_deferred);
-
-	string ext = enigma::image_get_format(filename);
-
-	LPDIRECT3DSURFACE9 pBackBuffer;
-	LPDIRECT3DSURFACE9 pDestBuffer;
-	d3ddev->GetBackBuffer(0, 0, D3DBACKBUFFER_TYPE_MONO, &pBackBuffer);
-	D3DSURFACE_DESC desc;
-	pBackBuffer->GetDesc(&desc);
-
-	d3ddev->CreateOffscreenPlainSurface( desc.Width, desc.Height, desc.Format, D3DPOOL_SYSTEMMEM, &pDestBuffer, NULL );
-	d3ddev->GetRenderTargetData(pBackBuffer, pDestBuffer);
-
-	D3DLOCKED_RECT rect;
-
-	pDestBuffer->LockRect(&rect, NULL, D3DLOCK_READONLY);
-	unsigned char* bitmap = static_cast<unsigned char*>(rect.pBits);
-	pDestBuffer->UnlockRect();
-
-	int ret = image_save(filename, bitmap, w, h, desc.Width, desc.Height, false);
-
-  pBackBuffer->Release();
-	pDestBuffer->Release();
-
-  return ret;
-}
-
-int background_create_from_screen(int x, int y, int w, int h, bool removeback, bool smooth, bool preload)
-{
-  return -1; //TODO: implement
-}
-
-int sprite_create_from_screen(int x, int y, int w, int h, bool removeback, bool smooth, bool preload, int xorig, int yorig) {
-  return -1; //TODO: implement
-}
-
-int sprite_create_from_screen(int x, int y, int w, int h, bool removeback, bool smooth, int xorig, int yorig) {
-  return sprite_create_from_screen(x, y, w, h, removeback, smooth, true, xorig, yorig);
-}
-
-void sprite_add_from_screen(int id, int x, int y, int w, int h, bool removeback, bool smooth) {
-
 }
 
 } // namespace enigma_user
