+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3screen.cpp
@@ -21,28 +21,12 @@
 #include "Graphics_Systems/OpenGL/GLscreen.h"
 #include "OpenGLHeaders.h"
 #include "Graphics_Systems/General/GSscreen.h"
-#include "Graphics_Systems/General/GSbackground.h"
-#include "Graphics_Systems/General/GSsprite.h"
-#include "Graphics_Systems/General/GStextures.h"
 #include "Graphics_Systems/General/GSprimitives.h"
-#include "Graphics_Systems/General/GSmatrix.h"
-#include "Graphics_Systems/General/GScolors.h"
 
-#include "Universal_System/image_formats.h"
-#include "Universal_System/nlpo2.h"
-#include "Universal_System/Resources/background_internal.h"
-#include "Universal_System/Resources/sprites_internal.h"
 #include "Universal_System/roomsystem.h"
 #include "Platforms/General/PFwindow.h"
 
-#include <string>
-#include <cstdio>
-
-//WE SHOULDN'T DO THIS! Don't specify namespaces like this - Harijs
-using namespace std;
-
 using namespace enigma;
-using namespace enigma_user;
 
 namespace enigma {
 
@@ -78,120 +62,6 @@ void screen_set_viewport(gs_scalar x, gs_scalar y, gs_scalar width, gs_scalar he
   //NOTE: OpenGL viewports are bottom left unlike Direct3D viewports which are top left
   glViewport(viewport_x, viewport_y, viewport_w, viewport_h);
   glScissor(viewport_x, viewport_y, viewport_w, viewport_h);
-}
-
-int screen_save(string filename) //Assumes native integers are little endian
-{
-  draw_batch_flush(batch_flush_deferred);
-
-  unsigned int w=window_get_width(),h=window_get_height(),sz=w*h;
-  string ext = enigma::image_get_format(filename);
-
-  unsigned char *rgbdata = new unsigned char[sz*4];
-  GLint prevFbo;
-  glGetIntegerv(GL_FRAMEBUFFER_BINDING, &prevFbo);
-  glPixelStorei(GL_PACK_ALIGNMENT, 1);
-  glBindFramebuffer(GL_READ_FRAMEBUFFER, 0);
-  glReadPixels(0,0,w,h, GL_BGRA, GL_UNSIGNED_BYTE, rgbdata);
-  glBindFramebuffer(GL_DRAW_FRAMEBUFFER, prevFbo);
-
-  int ret = image_save(filename, rgbdata, w, h, w, h, false);
-
-  delete[] rgbdata;
-  return ret;
-}
-
-int screen_save_part(string filename,unsigned x,unsigned y,unsigned w,unsigned h) //Assumes native integers are little endian
-{
-  draw_batch_flush(batch_flush_deferred);
-
-  unsigned sz = w*h;
-  string ext = enigma::image_get_format(filename);
-
-  unsigned char *rgbdata = new unsigned char[sz*4];
-  GLint prevFbo;
-  glGetIntegerv(GL_FRAMEBUFFER_BINDING, &prevFbo);
-  glPixelStorei(GL_PACK_ALIGNMENT, 1);
-  glBindFramebuffer(GL_READ_FRAMEBUFFER, 0);
-  glReadPixels(x,window_get_region_height_scaled()-h-y,w,h, GL_BGRA, GL_UNSIGNED_BYTE, rgbdata);
-  glBindFramebuffer(GL_DRAW_FRAMEBUFFER, prevFbo);
-
-  int ret = image_save(filename, rgbdata, w, h, w, h, false);
-
-  delete[] rgbdata;
-  return ret;
-}
-
-// The following three leave a bad taste in my mouth because they depend on views, which should be removable.
-// However, for now, they stay.
-
-int background_create_from_screen(int x, int y, int w, int h, bool removeback, bool smooth, bool preload) {
-	draw_batch_flush(batch_flush_deferred);
-
-  int full_width=enigma::nlpo2dc(w)+1, full_height=enigma::nlpo2dc(h)+1;
-	int prevFbo;
-	glGetIntegerv(GL_FRAMEBUFFER_BINDING, &prevFbo);
-	glBindFramebuffer(GL_READ_FRAMEBUFFER, 0);
-	int patchSize = full_width*full_height;
-	std::vector<unsigned char> rgbdata(4*patchSize);
-	glReadPixels(x, enigma_user::window_get_region_height_scaled()-h-y,w,h,GL_BGRA, GL_UNSIGNED_BYTE, &rgbdata[0]);
-	glBindFramebuffer(GL_DRAW_FRAMEBUFFER, prevFbo);
-
-	unsigned char* data = enigma::image_flip(&rgbdata[0], w, h, 4);
-
-	enigma::backgroundstructarray_reallocate();
-  int bckid=enigma::background_idmax;
-	enigma::background_new(bckid, w, h, &data[0], removeback, smooth, preload, false, 0, 0, 0, 0, 0, 0);
-  delete[] data;
-	rgbdata.clear();
-	enigma::background_idmax++;
-  return bckid;
-}
-
-int sprite_create_from_screen(int x, int y, int w, int h, bool removeback, bool smooth, bool preload, int xorig, int yorig) {
-  draw_batch_flush(batch_flush_deferred);
-
-  int full_width=enigma::nlpo2dc(w)+1, full_height=enigma::nlpo2dc(h)+1;
-	int prevFbo;
-	glGetIntegerv(GL_FRAMEBUFFER_BINDING, &prevFbo);
- 	glBindFramebuffer(GL_READ_FRAMEBUFFER, 0);
-	int patchSize = full_width*full_height;
-	std::vector<unsigned char> rgbdata(4*patchSize);
-	glReadPixels(x,enigma_user::window_get_region_height_scaled()-h-y,w,h,GL_BGRA, GL_UNSIGNED_BYTE, &rgbdata[0]);
-	glBindFramebuffer(GL_DRAW_FRAMEBUFFER, prevFbo);
-
-	unsigned char* data = enigma::image_flip(&rgbdata[0], w, h, 4);
-
-	enigma::spritestructarray_reallocate();
-  int sprid=enigma::sprite_idmax;
-  enigma::sprite_new_empty(sprid, 1, w, h, xorig, yorig, 0, h, 0, w, preload, smooth);
-	enigma::sprite_set_subimage(sprid, 0, w, h, &data[0], &data[0], enigma::ct_precise); //TODO: Support toggling of precise.
-  rgbdata.clear(); // Clear the temporary array
-	delete[] data;
-  return sprid;
-}
-
-int sprite_create_from_screen(int x, int y, int w, int h, bool removeback, bool smooth, int xorig, int yorig) {
-	return sprite_create_from_screen(x, y, w, h, removeback, smooth, true, xorig, yorig);
-}
-
-void sprite_add_from_screen(int id, int x, int y, int w, int h, bool removeback, bool smooth) {
-  draw_batch_flush(batch_flush_deferred);
-
-  int full_width=enigma::nlpo2dc(w)+1, full_height=enigma::nlpo2dc(h)+1;
-	int prevFbo;
-	glGetIntegerv(GL_FRAMEBUFFER_BINDING, &prevFbo);
- 	glBindFramebuffer(GL_READ_FRAMEBUFFER, 0);
-	int patchSize = full_width*full_height;
-	std::vector<unsigned char> rgbdata(4*patchSize);
-	glReadPixels(x, enigma_user::window_get_region_height_scaled()-h-y,w,h,GL_BGRA, GL_UNSIGNED_BYTE, &rgbdata[0]);
-	glBindFramebuffer(GL_DRAW_FRAMEBUFFER, prevFbo);
-
-	unsigned char* data = enigma::image_flip(&rgbdata[0], w, h, 4);
-
-	enigma::sprite_add_subimage(id, w, h, &data[0], &data[0], enigma::ct_precise); //TODO: Support toggling of precise.
-	delete[] data;
-	rgbdata.clear();
 }
 
 } // namespace enigma_user
